+++ src/screens/Calendar/Calendar.scss
@@ -16,7 +16,6 @@
     box-shadow: 0px 1px 3px 1px #1A237E26;
     background-color: white;
     
-
     .header-title {
       font-size: 32px;
       h2 {
@@ -36,7 +35,6 @@
     justify-content: flex-end;
     margin-right: 20px;
  
-
     .MuiPaper-root{
       position: fixed;
       bottom: 30px;
