+++ src/screens/Calendar/CalendarView.scss
@@ -1,5 +1,4 @@
 .calendar-container{
-  
 
   .fc .fc-today-button {
     text-transform: capitalize;
