+++ src/screens/Calendar/MeetingModal/styles/Datefields.scss
@@ -13,50 +13,17 @@
   margin-top: 10px;
 }
 
-.fields-wrapper {
-  // display: flex;
-  // flex-direction: column;
-  // grid-area: dateFields;
-  // justify-content:center;
-}
+.select-time-wrapper {
+  margin-left: 10px;
+  margin-right: 50px;
 
-.time-fields {
-  // align-items: center;
-  // display: flex;
-  // flex-direction: row;
-  // height: 40px;
-  // margin-left: 20px;
+ }
 
-  .time-icon {
-    // position: relative;
-    // right: 17px;
-    // width: 17px;
-  }
-
-  .select-time-wrapper {
-    margin-left: 10px;
-    margin-right: 50px;
-
-  }
-
-
-  .span-dash {
-    font-weight: 600;
-    padding: 10px;
-    letter-spacing: 2px;
-    font-size: 14px;
-  }
+.span-dash {
+  font-weight: 600;
+  padding: 10px;
+  letter-spacing: 2px;
+  font-size: 14px;
 }
 
 
-.timezone-wrapper {
-  // align-self: center;
-  // display: flex;
-  // flex-direction: row;
-  // justify-content: center;
-  // width: 100%;
-   p {
-    // font-size: 14px;
-    // margin-right: 19px;
-   }
-}
