+++ src/screens/Calendar/MeetingModal/styles/DiscardChangesModal.scss
@@ -5,7 +5,6 @@
   border-radius: 5px;
 }
 
-
 .discard-buttons-container {
   margin-top: 30px;
   justify-content: end;
