+++ src/screens/Calendar/MeetingModal/styles/MemberCheckbox.scss
@@ -5,9 +5,6 @@
   width: 70%;
   margin-left: -10px;
 
-
- 
-
   p {
     color: #616161;
     font-style: italic;
