+++ src/utils/redux/slices/calendarSlice.ts
@@ -49,12 +49,6 @@ const calendarSlice = createSlice({
       state.displayedEvent = action.payload
       state.modalDisplayStatus = 'display'
     },
-
-    deleteExistingEvent: (state, action) => {
-      const { eventId } = action.payload
-      state.convertedEvents.filter(event => event.eventId !== eventId)
-      delete state.eventMap[eventId]
-    },
   },
 })
 
@@ -79,7 +73,6 @@ export const {
   setDisplayedEvent,
   storeConvertedEvents,
   setModalDisplayStatus,
-  deleteExistingEvent,
 } = calendarSlice.actions
 
 export default calendarSlice.reducer
